+++ tests/testthat/test-restrict.R
@@ -1,12 +1,12 @@
-test_that("restrict", {
+test_that("restrict_fun", {
 
   g0 = function(x,y,z) {
     x + y + z + u
   }
 
-  g1 <- restrict_sub(g0, list(z=100))  
-  g2 <- restrict_sub(g1, list(y=10))
-  g3 <- restrict_sub(g2, list(x=1))
+  g1 <- restrict_fun_sub(g0, list(z=100))  
+  g2 <- restrict_fun_sub(g1, list(y=10))
+  g3 <- restrict_fun_sub(g2, list(x=1))
 
   u <- 1000
   expect_equal(g0(x=1, y=10, z=100), 1111)
@@ -18,9 +18,9 @@ test_that("restrict", {
     x + y + z + u
   }
 
-  ## f1 <- restrict(f0, list(z=200))  
-  ## f2 <- restrict(f1, list(y=20))
-  ## f3 <- restrict(f2, list(x=2))
+  ## f1 <- restrict_fun(f0, list(z=200))  
+  ## f2 <- restrict_fun(f1, list(y=20))
+  ## f3 <- restrict_fun(f2, list(x=2))
 
   
   ## u <- 2000